refactor(db): Split auto_commit_mutator.rs into directory module
