+++ crates/db/src/auto_commit_mutator/mod.rs
@@ -1,0 +1,102 @@
+//! Auto-committing document mutator for non-transactional mutations.
+//!
+//! This mutator wraps a database and automatically creates and commits
+//! a write transaction for each mutation operation. This enables mutations
+//! without explicit transaction management while still providing proper
+//! transactional semantics per operation.
+
+mod create;
+mod create_many;
+mod delete;
+pub(super) mod helpers;
+mod read;
+mod update;
+
+use async_trait::async_trait;
+use cid::Cid;
+use document::{DocID, Document};
+use events::{Message, Update};
+use query::mutator::{CreateResult, DeleteResult, DocMutator, UpdateResult};
+use std::sync::Arc;
+use storage::corekv::Store;
+use tracing::warn;
+
+use crate::block_builder::{write_collection_block, write_delete_block, write_document_blocks};
+use crate::collection::collection_short_id;
+use crate::database::DB;
+use crate::index_manager::IndexManager;
+use defra_core::encryption::{get_doc_encryption, get_encryption_config, store_doc_encryption};
+use defra_core::signing::get_signing_config;
+
+/// Document mutator that auto-commits transactions for each operation.
+///
+/// This is useful for mutations that don't need explicit transaction control.
+/// Each operation creates a new write transaction, performs the mutation,
+/// and commits (or discards on error).
+///
+/// # Transaction Semantics
+///
+/// Each mutation is atomic: it either succeeds entirely or fails without
+/// partial changes. However, multiple mutations are NOT atomic with respect
+/// to each other - if you need multiple operations to be atomic, use
+/// `DbDocMutator` with explicit transaction management instead.
+pub struct AutoCommitMutator<S: Store> {
+    db: Arc<DB<S>>,
+}
+
+impl<S: Store> AutoCommitMutator<S> {
+    /// Create a new auto-committing mutator wrapping the given database.
+    pub fn new(db: Arc<DB<S>>) -> Self {
+        Self { db }
+    }
+}
+
+#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
+#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
+impl<S: Store + 'static> DocMutator for AutoCommitMutator<S> {
+    async fn create(
+        &self,
+        collection_name: &str,
+        doc: Document,
+    ) -> query::error::Result<CreateResult> {
+        self.create_impl(collection_name, doc).await
+    }
+
+    async fn create_many(
+        &self,
+        collection_name: &str,
+        docs: Vec<Document>,
+    ) -> query::error::Result<Vec<CreateResult>> {
+        self.create_many_impl(collection_name, docs).await
+    }
+
+    async fn update(
+        &self,
+        collection_name: &str,
+        doc: Document,
+        modified_fields: std::collections::HashSet<String>,
+    ) -> query::error::Result<UpdateResult> {
+        self.update_impl(collection_name, doc, modified_fields)
+            .await
+    }
+
+    async fn delete(
+        &self,
+        collection_name: &str,
+        doc_id: &DocID,
+    ) -> query::error::Result<DeleteResult> {
+        self.delete_impl(collection_name, doc_id).await
+    }
+
+    async fn exists(&self, collection_name: &str, doc_id: &DocID) -> query::error::Result<bool> {
+        self.exists_impl(collection_name, doc_id).await
+    }
+
+    async fn get_for_update(
+        &self,
+        collection_name: &str,
+        doc_id: &DocID,
+    ) -> query::error::Result<Option<Document>> {
+        self.get_for_update_impl(collection_name, doc_id).await
+    }
+}
